시큐리티 오류 수정
