+++ .settings/org.eclipse.wst.jsdt.ui.superType.name
@@ -1,0 +1,1 @@
+Window
