+++ TeamProject/src/main/java/com/project/config/SecurityWebApplicationInitializer.java
@@ -1,0 +1,7 @@
+package com.project.config;
+
+import org.springframework.security.web.context.AbstractSecurityWebApplicationInitializer;
+
+public class SecurityWebApplicationInitializer extends AbstractSecurityWebApplicationInitializer {
+
+}
