+++ TeamProject/src/main/java/com/project/config/ServletConfig.java
@@ -9,7 +9,7 @@ import org.springframework.web.servlet.view.InternalResourceViewResolver;
 import org.springframework.web.servlet.view.JstlView;
 
 @EnableWebMvc
-@ComponentScan(basePackages = {"com.delivery.controller", "com.delivery.exception"})
+@ComponentScan(basePackages = {"com.project.controller", "com.project.exception"})
 public class ServletConfig implements WebMvcConfigurer {
 	@Override
 	public void configureViewResolvers(ViewResolverRegistry registry) {
