--- TeamProject/src/main/java/com/project/security/SecurityWebApplicationInitializer.java
@@ -1,7 +1,0 @@
-package com.project.security;
-
-import org.springframework.security.web.context.AbstractSecurityWebApplicationInitializer;
-
-public class SecurityWebApplicationInitializer extends AbstractSecurityWebApplicationInitializer {
-
-}
